requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
redis[hiredis]>=4.5.3

# http client
httpx==0.23.3

# database
SQLAlchemy==2.0.7
aiosqlite==0.18.0
asyncpg==0.27.0
alembic==1.10.2

# auth service
itsdangerous==2.1.2
requests-oauthlib==1.3.1
async-cache==1.1.1
PyJWT==2.6.0

# libvirt
libvirt-python==9.1.0

# job queue
arq==0.25.0
msgpack==1.0.5